refactor: mark shadcnui's carousel and gallery as legacy
